requirements.txt
wget==3.2
bs4==4.10.0
python-chess==0.31.4